agents/pydantic_ai_docs_bot_python/pyproject.toml
[project]
name = "pydantic-ai-docs-bot"
version = "0.1.0"
description = "Multi-agent documentation Q&A bot using PydanticAI and Temporal"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.1.0",
    "openai>=1.0.0",
    "temporalio>=1.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]
